#6 add empty code generator
The workflow can be set to generate a code generator including a main and its own workflow file (see comment above). The behavior of the generator is yet undefined. We have therefore the ability to generate text (i.e. source core) or a concrete Java model that implements the behavior specified by the input ECDAR file.