fix each traversal re-setting clock
Opposite to what ECDAR specifies, the clock was reset in the code for each single traversal. This has been removed and now is only executed if an update is performed on traversal. Of course, this is not quite right, but since we assume only a single clock, we are good for now.