fix issue where "patent" would never be signaled
The bug was simple: as we were always checking that the invariant for the target location would hold for the current time, the effect of the assignment happening at traversal time (i.e. resetting the clock to zero) would be neglected. Therefore, no transition would happen, leaving the system in an invalid state.